Rename ancestor utils
