+++ libs/designsystem/helpers/src/public_api.ts
@@ -1,6 +1,6 @@
 // re-export helpers from core, to preserve backwards compatability
 // of 'designsystem' library
-export * from './element-has-ancestor';
+export * from './element-ancestor-utils';
 export * from './color-helper';
 export * from './content-mutation-observer';
 export * from './design-token-helper';
